edited

+++ 0x01-variables_if_else_while/6-print_numberz.c
@@ -10,7 +10,7 @@
  */
 int main(void)
 {
-int expo = 0;
+long expo = 0;
 while (expo < 10)
 {
 putchar(expo);
